Move compliance ignore users from audit config to compliance config (#484)
* Move compliance ignore users from audit config to compliance config

(cherry picked from commit b7626eec5e269129c5b9455a5781e5a161b3cafe)
